Keyboard mapping now a loadable multiboot module
This wasn't really that big of a change, but a bunch of
things went into it. First of all this commit is not
final. There is a temporary change in `Cargo.toml` which
lets the project build with my temporary fork of Phil's
multiboot2 instead of his own until it merges. Secondly
we can, at the moment, *only handle one multiboot2 module
at a time* until multiboot2::TagIter becomes public or I
create a multiboot module iter, we can only access the
first one. This will be resolved. Finally we have switched
from using utf-8 chars in the keyboard mapping to plain-ol
ascii. This is because of how I'm building the mappings
with nasm. Eventually we should probably change this, but
for now it's fine.

To use the new features use make with the variable `kbmap`
defined with a kbmap you use. e.g. `make run kbmap=dvorak`
It defaults to us and we also support dvorak. To create
another keyboard mapping, $kbmap, use us as a template in
src/arch/x86_64/module/keyboard/us.asm and move it to
src/arch/x86_64/module/keyboard$kbmap.asm then if you run
`make kbmap=$kbmap` it should load it.

+++ src/interrupts/mod.rs
@@ -116,7 +116,7 @@ pub extern "C" fn rust_keyboard_interrupt_handler() {
         // then the top bit of it is set
         x if x & 0x80 == 0 => {
             kb.keys[x as usize] = true;
-            let mut byte = kb.kbmap[x as usize] as u8;
+            let mut byte = kb.kbmap[x as usize];
 
             // If either shift is pressed, make it
             // capital as long as it is alphabetic
